Simple Pythin Brute Force Script
If you have Cyber roam in college and want to exploit the users that
have naive passwords, then you can use this python script. Just make
sure you have required python libraries and the format of the ID matches
to your college. This is just for educational purpose, no dishonest work
is intended.